Extract Msf::DBManager::Import::Retina
MSP-11124
